Move new tests to new file
